backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.10.0
pydantic[email]==2.10.6
pydantic-settings==2.7.1
anthropic==0.40.0
reportlab==4.2.5
python-docx==1.1.2
pdfplumber==0.11.4
python-multipart==0.0.12
python-dotenv==1.0.1
